fix(api): hide DRAFT/SCHEDULED posts from public callers
The /api/blog list endpoint and /api/blog/[slug] detail endpoint
were returning DRAFT and SCHEDULED posts to anonymous callers (no
auth check, no status filter). Confirmed leak in production:

  curl -s 'https://richardwhudsonjr.com/api/blog?limit=20' | jq '.data | group_by(.status)'
  # SCHEDULED: 1, DRAFT: 1, PUBLISHED: 18

Anyone with knowledge of (or willing to guess) a draft slug could
read full unpublished content. The /blog/[slug] page render filters
correctly (renders not-found for non-PUBLISHED), so end users on the
website never saw drafts — but the API was wide open.

Fixes:

1. New `src/lib/api-admin-auth.ts` — `isAdminRequest()` extracted
   from /api/seed/route.ts, now shared. Constant-time bearer check
   against ADMIN_API_TOKEN.

2. /api/blog GET — public callers default to `status='PUBLISHED' AND
   deletedAt IS NULL`. Admin-token holders see everything (or pass
   `?status=DRAFT` etc. for explicit selection). buildBlogWhereClause
   now also adds `deletedAt IS NULL` whenever `published=true` is set,
   so future soft-deletes propagate immediately.

3. /api/blog/[slug] GET — anonymous callers requesting a non-published
   or soft-deleted slug get HTTP 404 (same response as a missing slug,
   so they can't differentiate). Admin tokens bypass. View count is
   only incremented on actually-public reads. Cache-Control becomes
   `private, no-store` for admin-only views to keep CDN tiers from
   accidentally caching unpublished content.

4. /api/seed — now imports the shared isAdminRequest instead of the
   inlined isAuthorized helper. Behavior unchanged.

Verified: typecheck clean, lint clean, 181 tests pass.

[hudsor01]
